[HEVC FEI] Fixed broken HEVC FEI
HEVC FEI has been initializing with VAEntrypointEncSlice
which is definitely wrong
