+++ _studio/mfx_lib/encode_hw/h265/include/mfx_h265_fei_encode_vaapi.h
@@ -1,15 +1,15 @@
 // Copyright (c) 2017-2018 Intel Corporation
-// 
+//
 // Permission is hereby granted, free of charge, to any person obtaining a copy
 // of this software and associated documentation files (the "Software"), to deal
 // in the Software without restriction, including without limitation the rights
 // to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
 // copies of the Software, and to permit persons to whom the Software is
 // furnished to do so, subject to the following conditions:
-// 
+//
 // The above copyright notice and this permission notice shall be included in all
 // copies or substantial portions of the Software.
-// 
+//
 // THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
 // IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
 // FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
@@ -76,14 +76,11 @@ namespace MfxHwH265FeiEncode
 
         virtual mfxStatus PostQueryExtraStage(MfxHwH265Encode::Task const & task, mfxU32 codedStatus) override;
 
-        virtual VAEntrypoint GetVAEntryPoint() override
-        {
-            return VAEntrypointFEI;
-        }
-
         virtual mfxStatus ConfigureExtraVAattribs(std::vector<VAConfigAttrib> & attrib) override;
 
         virtual mfxStatus CheckExtraVAattribs(std::vector<VAConfigAttrib> & attrib) override;
+
+        virtual MfxHwH265Encode::VAParameters GetVaParams(const GUID & guid) override;
     };
 }
